add bool check #2
ketinggalan long int-nya hahahaha

+++ evaluasiAritmatika.cpp
@@ -16,7 +16,7 @@ int main(){
 bool check(char ch, vector<char> range){
     vector<char>::iterator it = find(range.begin(),range.end(),ch);;
 
-    return it - range.begin() < range.size();
+    return it - range.begin() < (long int) range.size();
 }
 
 string removeWhitespace(string str){
